cleanup failure logic, remove constants, fix tests
