+++ test/fixtures/custom/invalid-failure-string.js
@@ -4,18 +4,18 @@ import { superstruct } from '../../..'
 const struct = superstruct({
   types: {
     email: v => {
-      if (!isEmail(v)) return { message: `not_email` }
+      if (!isEmail(v)) return `not_email`
     },
   },
 })
 
-export const Struct = struct('email?')
+export const Struct = struct('email')
 
-export const data = 'optional_invalid'
+export const data = 'invalid'
 
 export const error = {
   path: [],
-  value: 'optional_invalid',
-  type: 'email | undefined',
+  value: 'invalid',
+  type: 'email',
   reason: 'not_email',
 }
