+++ test/fixtures/custom/optional-invalid-failure-object.js
@@ -4,7 +4,7 @@ import { superstruct } from '../../..'
 const struct = superstruct({
   types: {
     email: v => {
-      if (!isEmail(v)) return { message: `not_email` }
+      if (!isEmail(v)) return { reason: `not_email`, path: ['email'] }
     },
   },
 })
@@ -14,7 +14,7 @@ export const Struct = struct('email?')
 export const data = 'optional_invalid'
 
 export const error = {
-  path: [],
+  path: ['email'],
   value: 'optional_invalid',
   type: 'email | undefined',
   reason: 'not_email',
