+++ test/fixtures/custom/optional-invalid-failure-string.js
@@ -4,18 +4,18 @@ import { superstruct } from '../../..'
 const struct = superstruct({
   types: {
     email: v => {
-      if (!isEmail(v)) return { message: `not_email` }
+      if (!isEmail(v)) return `not_email`
     },
   },
 })
 
-export const Struct = struct('email')
+export const Struct = struct('email?')
 
-export const data = 'invalid'
+export const data = 'optional_invalid'
 
 export const error = {
   path: [],
-  value: 'invalid',
-  type: 'email',
+  value: 'optional_invalid',
+  type: 'email | undefined',
   reason: 'not_email',
 }
